Handle S3 API failures that returns 200 OK
The SDK relies on HTTP status codes to determine whether or not the
operation succeeded. However, in certain APIs such as
CompleteMultipartUpload, the API may return 200 OK with an error in the
response body. In those cases, the response body needs to be inspected
for errors. This fix only checks response bodies of non-GET requests
that return XML as the response type.